@@ -1,73 +1,51 @@
-import React, { Component } from "react";
-import SimpleStorageContract from "./contracts/SimpleStorage.json";
-import getWeb3 from "./getWeb3";
-
 import "./App.css";
+import HomePage from "./component/HomePage";
+import ProjectDetail from "./component/ProjectDetail";
+import DonateStatus from "./component/DonateStatus";
 
-class App extends Component {
-  state = { storageValue: 0, web3: null, accounts: null, contract: null };
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import AddNewProject from "./component/AddNewProject";
+import NavBar from "./component/NavBar";
+import { VscGithubInverted } from "react-icons/vsc";
+import TimeLine from "./component/TimeLine";
 
-  componentDidMount = async () => {
-    try {
-      // Get network provider and web3 instance.
-      const web3 = await getWeb3();
-
-      // Use web3 to get the user's accounts.
-      const accounts = await web3.eth.getAccounts();
-
-      // Get the contract instance.
-      const networkId = await web3.eth.net.getId();
-      const deployedNetwork = SimpleStorageContract.networks[networkId];
-      const instance = new web3.eth.Contract(
-        SimpleStorageContract.abi,
-        deployedNetwork && deployedNetwork.address,
-      );
-
-      // Set web3, accounts, and contract to the state, and then proceed with an
-      // example of interacting with the contract's methods.
-      this.setState({ web3, accounts, contract: instance }, this.runExample);
-    } catch (error) {
-      // Catch any errors for any of the above operations.
-      alert(
-        `Failed to load web3, accounts, or contract. Check console for details.`,
-      );
-      console.error(error);
-    }
-  };
-
-  runExample = async () => {
-    const { accounts, contract } = this.state;
-
-    // Stores a given value, 5 by default.
-    await contract.methods.set(5).send({ from: accounts[0] });
-
-    // Get the value from the contract to prove it worked.
-    const response = await contract.methods.get().call();
-
-    // Update state with the result.
-    this.setState({ storageValue: response });
-  };
-
-  render() {
-    if (!this.state.web3) {
-      return <div>Loading Web3, accounts, and contract...</div>;
-    }
-    return (
-      <div className="App">
-        <h1>Good to Go!</h1>
-        <p>Your Truffle Box is installed and ready.</p>
-        <h2>Smart Contract Example</h2>
-        <p>
-          If your contracts compiled and migrated successfully, below will show
-          a stored value of 5 (by default).
-        </p>
-        <p>
-          Try changing the value stored on <strong>line 42</strong> of App.js.
-        </p>
-        <div>The stored value is: {this.state.storageValue}</div>
-      </div>
-    );
-  }
+function App() {
+  return (
+    <Router>
+      <header>
+        <NavBar />
+      </header>
+      <body className="body">
+        <Routes>
+          <Route exact path="/" element={<HomePage />} />
+          <Route exact path="/detail/:no" element={<ProjectDetail />} />
+          <Route exact path="/t" element={<TimeLine />} />
+          <Route exact path="/manager" element={<AddNewProject />} />
+          <Route exact path="/detail/:no/status" element={<DonateStatus />} />
+        </Routes>
+      </body>
+      <footer className="footer">
+        <div className="footer-bottom">
+          <div className="footer-content">
+            <h3>project title</h3>
+            <p>
+              Lorem Ipsum is simply dummy text of the printing and typesetting
+              iLorem Ipsum is simply dummy text of the printing and typesetting
+              iLorem Ipsum is simply dummy text of the printing and typesetting
+              i
+            </p>
+            <ul className="socials">
+              <li>
+                <a href="https://github.com/SO-TY">
+                  <VscGithubInverted className="icon" />
+                </a>
+              </li>
+            </ul>
+          </div>
+        </div>
+      </footer>
+    </Router>
+  );
 }
 
 export default App;
